Add files via upload
I had a little spare time to clean it up a little and added a class and function to make it oop. If anybody has any input on how to make the a and b two different function but still pass the value of b to the second function please let me know.